[fix] correct flip x/y axes reading igor ibw
scifireaders does not properly load the scan data. The data is read
originally in the same manner as gwyddion, which causes the origin to be
misplaced.

In the Asylum UI, the data is displayed as:
  (y)
   ^
   |
   |
   |
   |
   x-------> (x)
(0, 0)

In Gwyddion, the data is displayed as:
(0, 0)
   x-------> (x)
   |
   |
   |
   |
   V
  (y)

In both cases, the actual data is the same! Thus, Gwyddion incorrectly
sets the origin and axes.

In an attempt to fix this, the IgorIBWReader performs np.rot90(m, 3),
i.e. rotates by 90 \deg 3x counter-clockwise. This *mostly* works,
except that it swaps the axes!
We end up with:
  (x)
   ^
   |
   |
   |
   |
   x-------> (y)
(0, 0)

The more 'proper' fix would be to simple perform an np.flip(m, 0) rather
than these operations. This is what is done in this CL.

Note also that np.rot90 is used in spm/afm/mdt.py and spm/afm/wxsm.py. I
did not include those in this CL because I do not have these instruments
and thus could not confirm the same mistake is being made.